feat: get funnel steps graph data from API + move logic to custom hook
