+++ frontend/src/constants/reactQueryKeys.ts
@@ -74,4 +74,5 @@ export const REACT_QUERY_KEY = {
 	GET_FUNNEL_OVERVIEW: 'GET_FUNNEL_OVERVIEW',
 	GET_FUNNEL_SLOW_TRACES: 'GET_FUNNEL_SLOW_TRACES',
 	GET_FUNNEL_ERROR_TRACES: 'GET_FUNNEL_ERROR_TRACES',
+	GET_FUNNEL_STEPS_GRAPH_DATA: 'GET_FUNNEL_STEPS_GRAPH_DATA',
 } as const;
