+++ frontend/src/hooks/TracesFunnels/useFunnels.tsx
@@ -6,11 +6,13 @@ import {
 	ErrorTracesPayload,
 	FunnelOverviewPayload,
 	FunnelOverviewResponse,
+	FunnelStepsResponse,
 	getFunnelById,
 	getFunnelErrorTraces,
 	getFunnelOverview,
 	getFunnelsList,
 	getFunnelSlowTraces,
+	getFunnelSteps,
 	renameFunnel,
 	saveFunnelDescription,
 	SlowTraceData,
@@ -137,6 +139,7 @@ export const useValidateFunnelSteps = ({
 			),
 		queryKey: [REACT_QUERY_KEY.VALIDATE_FUNNEL_STEPS, funnelId, selectedTime],
 		enabled: !!funnelId && !!selectedTime && !!startTime && !!endTime,
+		staleTime: 1000 * 60 * 5,
 	});
 
 export const useUpdateFunnelStepDetails = ({
@@ -211,3 +214,29 @@ export const useFunnelErrorTraces = (
 		enabled: !!funnelId && validTracesCount > 0,
 	});
 };
+
+export function useFunnelStepsGraphData(
+	funnelId: string,
+): UseQueryResult<SuccessResponse<FunnelStepsResponse> | ErrorResponse, Error> {
+	const {
+		startTime,
+		endTime,
+		selectedTime,
+		validTracesCount,
+	} = useFunnelContext();
+
+	return useQuery({
+		queryFn: ({ signal }) =>
+			getFunnelSteps(
+				funnelId,
+				{ start_time: startTime, end_time: endTime },
+				signal,
+			),
+		queryKey: [
+			REACT_QUERY_KEY.GET_FUNNEL_STEPS_GRAPH_DATA,
+			funnelId,
+			selectedTime,
+		],
+		enabled: !!funnelId && validTracesCount > 0,
+	});
+}
